Move MapViewModel to maps module
