+++ osmdroid/src/main/java/org/odk/collect/osmdroid/OsmDroidMapFragment.java
@@ -49,7 +49,9 @@ import org.odk.collect.maps.MapConfigurator;
 import org.odk.collect.maps.MapFragment;
 import org.odk.collect.maps.MapFragmentDelegate;
 import org.odk.collect.maps.MapPoint;
+import org.odk.collect.maps.MapViewModel;
 import org.odk.collect.maps.PolygonDescription;
+import org.odk.collect.maps.Zoom;
 import org.odk.collect.maps.layers.MapFragmentReferenceLayerUtils;
 import org.odk.collect.maps.layers.ReferenceLayerRepository;
 import org.odk.collect.maps.markers.MarkerDescription;
